refactor(archestra-bench): dedup config helpers + extract chat_stream
- delete dead deserialize_table; reuse core is_slug + split_names
- move chat-stream types, apply_chat_event, SSE parser into runner/src/chat_stream.rs
- behavior-preserving relocation + dedup; net -16 lines
